fix: Remove task_role_arns from ecs.5 check (#1307)
* fix: remove task_role_arns from ecs.5 check

In some cases, these can be NULL. Since they are not required,
removing them means that there will be fewer latest_revisions,
which means that we see the expected behaviour.

Expected behaviour is that, following a sync, if the customer
has fixed the filesystem, we want to see this reflected in
more passing checks.

Consider the case where a container has a NULL task role ARN,
and as part of the customer's fix the filesystem of the task definition
is made readonly and the ARN is set. Prior to this PR, the
container would still be joined to the old task definition
unless its task role ARN were also changed.

After this fix, even if the task role ARN changes, we still
see the expected result (make the filesystem read only, and
the test passes on the next sync).

* chore: Update readme